move tests to proper directory
